src/back/requirements.txt
elasticsearch==8.13.0
elasticsearch-dsl==8.13.0
farm-haystack==1.24.1
fastapi==0.110.2
nmslib@git+https://github.com/nmslib/nmslib.git#egg=nmslib&subdirectory=python_bindings
numpy==1.26.4
pandas==2.2.2
pybind11==2.11.1
redis[hiredis]==5.0.4
sentence-transformers==2.7.0
spacy==3.7.4
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_scibert-0.5.4.tar.gz
spacyfishing==0.1.8
scispacy==0.5.4
uvicorn[standard]==0.29.0
websockets==12.0
nltk==3.7